@@ -36,8 +36,11 @@
     "mobx": "^2.1.1",
     "mobx-react": "^3.0.3",
     "mobx-react-devtools": "^4.0.2",
-    "react": "^15.0.1",
-    "react-dom": "^15.0.1",
     "webpack": "^1.12.11"
+  },
+  "peerDependencies": {
+    "prop-types": "^15.6.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
   }
 }
